% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_date_range_all.R
\name{prism_date_range_all}
\alias{prism_date_range_all}
\title{PRISM weather data at a given point within California for a given date
range}
\usage{
prism_date_range_all(con, lat, long, from_date, to_date)
}
\arguments{
\item{con}{A database connection}

\item{lat}{The latitude of the point of interest as a number. Must be within CA.}

\item{long}{The longitue of the point of interest as a number. Must be within CA.}

\item{from_date}{The start date of interest as "YYYY-MM-DD" character format
(data not available before 2009 currently)}

\item{to_date}{The end date of interest as "YYYY-MM-DD" character format
(data not available before 2009 currently)}
}
\value{
a data.frame object with daily values for the given date range
}
\description{
\code{prism_date_range_all} returns the daily values of the PRISM
tiles that intersect with the given point over the given date range for the
given the data type.
}
\details{
This function is written to work with the Grain Cropping Systems database at
UC Davis. A DBI connection to the database labeled "con" must be present in
the Global Environment. Example below:

con <- DBI::dbConnect(
RPostgres::Postgres(),
dbname = "mldatadb",
host = "169.237.215.4",
port = 5432,
user = Sys.getenv("userid"),
password = Sys.getenv("pwd"),
sslmode = 'require',
options="-c search_path=grain,public"
)
}
\examples{
prism_date_range_all(con = con, lat = 38.533867, long = -121.771598, from_date = "2019-10-01", to_date = "2019-12-01")
}
