% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_historical_season_all_steps.R
\name{prism_historical_season_all}
\alias{prism_historical_season_all}
\title{historical PRISM data at a given point within California as a daily average of
the 10 years previous to the current date starting at the date given and
ending at June 30 (for wheat seasonality)}
\usage{
prism_historical_season_all(
  con,
  lat,
  long,
  current_start_date = "2019-10-01",
  end_date
)
}
\arguments{
\item{con}{A database connection}

\item{lat}{The latitude of the point of interest as a number. Must be within CA.}

\item{long}{The longitue of the point of interest as a number. Must be within CA.}

\item{current_start_date}{user defined start date}

\item{end_date}{seasonal end date given the region and the user defined end_date}
}
\value{
a data.frame object with average historical daily values for the
chosen variable from the given date to the June 30th output as month, day,
and amount
}
\description{
\code{prism_historical_season_all} returns the average daily values of the PRISM
tiles that intersect with the given point over the previous 10 years starting
with the given month and day until June 30th given the data type.
}
\details{
This function is written to work with the Grain Cropping Systems database at
UC Davis. A DBI connection to the database labeled "con" must be present in
the Global Environment. Example below:

con <- DBI::dbConnect(
RPostgres::Postgres(),
dbname = "databasename",
host = "IP",
port = port,
user = Sys.getenv("userid"),
password = Sys.getenv("pwd"),
sslmode = 'require',
options="-c search_path=path_names"
)
}
\examples{
prism_historical_season_all(con = con, lat = 38.533867, long = -121.771598, current_start_date = "2019-10-01", end_date = "2020-06-30")
}
